refactor(top_cards): split TopCardsFrame into mixin package
Convert widgets/frames/top_cards.py (133 LOC) into a package following
the match_history/timer_alert layout:

- __init__.py: public API (TopCardsFrame) plus re-exports of the
  TOP_CARDS_EXCLUDED_FORMATS / TOP_CARDS_FORMAT_OPTIONS constants so
  existing ``from widgets.frames.top_cards import TopCardsFrame`` (and
  any future constant imports) keep working unchanged
- frame.py: sys.path bootstrap, constant definitions,
  TopCardsFrame.__init__, _build_ui, _stylize_button
- handlers.py: TopCardsHandlersMixin -- on_format_change, refresh_data
  (pulls summary + top cards from FormatCardPoolService and fills the
  ListCtrl), on_close
- properties.py: TopCardsPropertiesMixin -- translation helper (_t)

Composition becomes:
  class TopCardsFrame(HandlersMixin, PropertiesMixin, wx.Frame): ...

Existing imports in widgets/handlers/app_event_handlers.py,
widgets/frames/app_frame.py, and the lazy-export map in
widgets/frames/__init__.py continue to resolve TopCardsFrame via the
package __init__, so no call sites change.

pyproject.toml's E402 per-file-ignore now points at the new
widgets/frames/top_cards/frame.py where the sys.path bootstrap lives
(replacing the previous ``# ruff: noqa: E402`` file-level directive).

Refs #427 (second pass, partial).
