+++ widgets/frames/top_cards/__init__.py
@@ -1,0 +1,11 @@
+"""Top Cards viewer package."""
+
+from __future__ import annotations
+
+from widgets.frames.top_cards.frame import (
+    TOP_CARDS_EXCLUDED_FORMATS,
+    TOP_CARDS_FORMAT_OPTIONS,
+    TopCardsFrame,
+)
+
+__all__ = ["TOP_CARDS_EXCLUDED_FORMATS", "TOP_CARDS_FORMAT_OPTIONS", "TopCardsFrame"]
